Move InMemoryLogRecordExporter to logs-testing
