+++ sdk/logs-testing/src/main/java/io/opentelemetry/sdk/testing/exporter/InMemoryLogRecordExporter.java
@@ -3,10 +3,11 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.sdk.logs.export;
+package io.opentelemetry.sdk.testing.exporter;
 
 import io.opentelemetry.sdk.common.CompletableResultCode;
 import io.opentelemetry.sdk.logs.data.LogRecordData;
+import io.opentelemetry.sdk.logs.export.LogRecordExporter;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
