+++ sdk/logs-testing/src/test/java/io/opentelemetry/sdk/testing/exporter/InMemoryLogRecordExporterTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.sdk.logs.export;
+package io.opentelemetry.sdk.testing.exporter;
 
 import static io.opentelemetry.api.logs.Severity.DEBUG;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -11,6 +11,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import io.opentelemetry.api.logs.Logger;
 import io.opentelemetry.sdk.logs.SdkLoggerProvider;
 import io.opentelemetry.sdk.logs.data.LogRecordData;
+import io.opentelemetry.sdk.logs.export.SimpleLogRecordProcessor;
 import io.opentelemetry.sdk.testing.assertj.LogAssertions;
 import io.opentelemetry.sdk.testing.logs.TestLogRecordData;
 import java.util.Collections;
